+++ include/dn/mem/virtual.h
@@ -1,7 +1,7 @@
 #pragma once
 
-uint32_t DnMemVirtualPageSize();
-void* DnMemVirtualReserve(size_t size);
-bool DnMemVirtualCommit(void* ptr, size_t size);
-bool DnMemVirtualDecommit(void* ptr, size_t size);
+u32 DnMemVirtualPageSize();
+void* DnMemVirtualReserve(u64 size);
+bool DnMemVirtualCommit(void* ptr, u64 size);
+bool DnMemVirtualDecommit(void* ptr, u64 size);
 bool DnMemVirtualRelease(void* ptr);
